+++ packages/app/src/pages/endpoints/[endpointId]/_content/_hooks/useDescendants.ts
@@ -3,21 +3,20 @@ import { useMemo } from 'react';
 import { Document, Info, Method, OperationId } from '$types/oas';
 import {
   getContentBaseOperationResponseKeys,
-  getPathItem,
   getRequest,
   getRequestParameterKeys,
 } from '$utils/oas';
 
-export type UseRelatedReturn = {
+export type UseDescendantsReturn = {
   operationIds: OperationId[];
 };
-// Related Opeartions are
-// [1] operations whose path is the same as the base operation's one but method is different.
-// [2] action operations whose request payload key set doesn't contain one of base operation's response payload keys.
-const useRelated = function (
+// Descendant Opeartions are
+// [1] operations whose path includes base operation's one at the head.(i.e. path matches `/${baseOperation's path}/xxx/yyy`)
+// [2] action operations whose request payload key set contains one of base operation's response payload keys.
+const useDescendants = function (
   document: Document,
   content: Info['x-pages'][number]['contents'][number]
-): UseRelatedReturn {
+): UseDescendantsReturn {
   const operationIds = useMemo<OperationId[]>(function () {
     const _operationIds: OperationId[] = [];
     const baseRequest = getRequest(document, {
@@ -26,33 +25,31 @@ const useRelated = function (
     if (!baseRequest) {
       return _operationIds;
     }
-    const basePathItem = getPathItem(document, baseRequest.path);
-    if (!basePathItem) {
-      return _operationIds;
-    }
 
     // Add operations under the rule of [1] described above.
-    // TODO: method一覧の扱い改善。
-    (
-      [
-        'get',
-        'put',
-        'post',
-        'delete',
-        'options',
-        'head',
-        'patch',
-        'trace',
-      ] as Method[]
-    ).forEach(function (method) {
-      if (method === baseRequest.method) {
-        return;
-      }
-      const operationId = basePathItem[method]?.operationId;
-      if (!operationId) {
-        return;
-      }
-      _operationIds.push(operationId);
+    const pathItems = _.filter(document.paths, function (_, path) {
+      return path !== baseRequest.path && path.indexOf(baseRequest.path) === 0;
+    });
+    _.each(pathItems, function (pathItem) {
+      // TODO: method一覧の扱い改善。
+      (
+        [
+          'get',
+          'put',
+          'post',
+          'delete',
+          'options',
+          'head',
+          'patch',
+          'trace',
+        ] as Method[]
+      ).forEach(function (method) {
+        const operationId = pathItem[method]?.operationId;
+        if (!operationId) {
+          return;
+        }
+        _operationIds.push(operationId);
+      });
     });
 
     // Add operations under the rule of [2] described above.
@@ -68,7 +65,7 @@ const useRelated = function (
         );
         if (
           _.intersection(baseOperationResponseKeys, actionOperationRequestKeys)
-            .length === 0
+            .length
         ) {
           _operationIds.push(action.operationId);
         }
@@ -82,4 +79,4 @@ const useRelated = function (
     operationIds,
   };
 };
-export default useRelated;
+export default useDescendants;
